% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstimWithFrailty}
\alias{EstimWithFrailty}
\title{EM Algorithm for Frailty Models}
\usage{
EstimWithFrailty(
  s,
  n,
  nvar,
  k,
  nk,
  tau,
  caltimes,
  gaptimes,
  censored,
  intercepts,
  slopes,
  lastperrep,
  perrepind,
  effagebegin,
  effage,
  ndiseff,
  diseff,
  cov,
  alphaSeed,
  betaSeed,
  xi,
  Z,
  offset,
  rhoFunc,
  tol,
  maxiter,
  maxXi,
  estimEnd,
  control,
  loglikEnd
)
}
\description{
Estimates parameters for a recurrent event model with gamma frailty.
}
