% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstLambSurv}
\alias{EstLambSurv}
\title{Non-parametric Hazard and Survival Estimation}
\usage{
EstLambSurv(
  s,
  n,
  nvar,
  k,
  nk,
  tau,
  caltimes,
  gaptimes,
  censored,
  intercepts,
  slopes,
  lastperrep,
  perrepind,
  effagebegin,
  effage,
  ndiseff,
  diseff,
  cov,
  alpha,
  beta,
  Z,
  offset,
  rhoFunc,
  lambda,
  deltalambda,
  surv
)
}
\arguments{
\item{s}{Current calendar time.}

\item{n}{Number of subjects.}

\item{nvar}{Number of covariates.}

\item{k}{Vector with number of observations per subject.}

\item{nk}{Total number of observations.}

\item{tau}{Truncation times.}

\item{caltimes}{Calendar times of events.}

\item{gaptimes}{Gap times between events.}

\item{censored}{Censoring indicators.}

\item{intercepts}{Covariate intercepts.}

\item{slopes}{Covariate slopes.}

\item{lastperrep}{Last periodic report time.}

\item{perrepind}{Periodic report indicator.}

\item{effagebegin}{Effective age at start.}

\item{effage}{Effective age at event.}

\item{ndiseff}{Number of unique event times.}

\item{diseff}{Sorted unique event times.}

\item{cov}{Covariate matrix.}

\item{alpha}{Alpha parameter (intensity scaling).}

\item{beta}{Beta coefficients vector.}

\item{Z}{Vector of frailties.}

\item{offset}{Offset vector.}

\item{rhoFunc}{Model type indicator.}
}
\value{
A List containing lambdafunc, deltalambdafunc, and survfunc.
}
\description{
Computes the baseline cumulative hazard (Nelson-Aalen) and 
survival (Kaplan-Meier) functions using the at-risk process.
}
