% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survr.R
\name{is.Survr}
\alias{is.Survr}
\title{Test for a survival recurrent object}
\usage{
is.Survr(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{Survr}.
}
\description{
Test for a survival recurrent object
}
\examples{
is.Survr(Survr(id = c(1, 1), time = c(5, 3), event = c(1, 0)))
}
\seealso{
\code{\link[=Survr]{Survr()}}
}
