% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Jacknife2}
\alias{Jacknife2}
\title{Jacknife estimation (With Frailties)}
\usage{
Jacknife2(
  s,
  n,
  nvar,
  k,
  nk,
  tau,
  caltimes,
  gaptimes,
  censored,
  intercepts,
  slopes,
  lastperrep,
  perrepind,
  effagebegin,
  effage,
  ndiseff,
  diseff,
  cov,
  alphaSeed,
  betaSeed,
  xi,
  Z,
  offset,
  rhoFunc,
  tol,
  maxiter,
  maxXi
)
}
\arguments{
\item{s}{Scale parameter.}

\item{n}{Number of subjects.}

\item{nvar}{Number of covariates.}

\item{k}{Integer vector with the number of events per subject.}

\item{nk}{Total number of observations.}

\item{tau}{Vector of event/censoring times.}

\item{caltimes}{Vector of calendar times.}

\item{gaptimes}{Vector of gap times.}

\item{censored}{Censoring indicator vector.}

\item{intercepts}{Vector of intercepts.}

\item{slopes}{Vector of slopes.}

\item{lastperrep}{Vector of last perfect repair times.}

\item{perrepind}{Indicator of perfect repair.}

\item{effagebegin}{Effective age start.}

\item{effage}{Effective age end.}

\item{ndiseff}{Number of distinct effective ages.}

\item{diseff}{Vector of distinct effective ages.}

\item{cov}{Matrix of covariates.}

\item{alphaSeed}{Initial alpha.}

\item{betaSeed}{Initial betas.}

\item{xi}{Initial xi (frailty variance).}

\item{Z}{Initial frailty vector.}

\item{offset}{Vector of offsets.}

\item{rhoFunc}{Type of rho function.}

\item{tol}{Tolerance.}

\item{maxiter}{Max iterations.}

\item{maxXi}{Optimization method for xi.}
}
\value{
A matrix of size (n x nvar+2) containing parameter estimates (alpha, betas, xi) for each iteration.
}
\description{
Internal C++ function to perform Jackknife leave-one-out cross-validation for models with frailty.
}
