% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma}
\alias{lymphoma}
\title{Indolent non-Hodgkin's lymphomas}
\format{
A data frame with 112 rows and 9 columns:
\describe{
\item{id}{identifier of each subject, repeated for each recurrence}
\item{time}{inter-occurrence or censoring time}
\item{event}{censoring status: 1 for each relapse, 0 for the last
(censored) record of each subject}
\item{enum}{which relapse}
\item{delay}{delay between first symptom and date of first treatment}
\item{age}{age at diagnosis}
\item{sex}{gender: 1: Males, 2: Females}
\item{distrib}{lesions involved at diagnosis: 0: Single, 1: Localized,
2: More than one nodal site, 3: Generalized}
\item{effage}{response achieved after the treatment at each relapse:
\code{"CR"} complete remission, \code{"PR"} partial remission, \code{"SD"} stable
disease or null response. Used as the \code{cancer} argument of
\code{\link[=gcmrec]{gcmrec()}}}
}
}
\source{
González, J.R., Peña, E. and Slate, E. (2005). Modelling
intervention effects after cancer relapses. \emph{Statistics in Medicine},
24(24), 3959--3975.

Servitje, O., Gallardo, F., Estrach, T. et al. (2002). Primary
cutaneous marginal zone B-cell lymphoma. \emph{British Journal of
Dermatology}, 147, 1147--1158.
}
\usage{
lymphoma
}
\description{
Cancer relapse times after first treatment in patients diagnosed with
low grade lymphoma.
}
\examples{
data(lymphoma)
head(lymphoma)
}
\keyword{datasets}
