% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newtraph}
\alias{newtraph}
\title{Master Newton-Raphson Optimizer}
\usage{
newtraph(
  s,
  n,
  nvar,
  k,
  nk,
  tau,
  caltimes,
  gaptimes,
  censored,
  intercepts,
  slopes,
  lastperrep,
  perrepind,
  effagebegin,
  effage,
  cov,
  alphaSeed,
  betaSeed,
  Z,
  offset,
  rhoFunc,
  ns,
  tol,
  maxiter,
  loglik,
  estiEnd,
  info,
  searchBoth,
  kkiter
)
}
\arguments{
\item{s}{Current time point.}

\item{n}{Number of subjects.}

\item{nvar}{Number of covariates.}

\item{k}{Number of observations per subject.}

\item{nk}{Total number of observations across all subjects.}

\item{tau}{Vector of times.}

\item{caltimes}{Calendar times.}

\item{gaptimes}{Gap times.}

\item{censored}{Censoring indicator.}

\item{intercepts}{Intercepts for effective age.}

\item{slopes}{Slopes for effective age.}

\item{lastperrep}{Last perfect repair.}

\item{perrepind}{Perfect repair indicator.}

\item{effagebegin}{Effective age at start.}

\item{effage}{Effective age at end.}

\item{cov}{Matrix of covariates.}

\item{alphaSeed}{Initial value for alpha.}

\item{betaSeed}{Initial vector for betas.}

\item{Z}{Frailty vector.}

\item{offset}{Offset vector.}

\item{rhoFunc}{Type of rho function.}

\item{ns}{Internal subject vector.}

\item{tol}{Convergence tolerance.}

\item{maxiter}{Maximum number of iterations.}

\item{loglik}{Output: Final log-likelihood.}

\item{estiEnd}{Output: Vector of final estimates (alpha + betas).}

\item{info}{Output: Final information matrix.}

\item{searchBoth}{Output: Status of the final simultaneous optimization.}

\item{kkiter}{Output: Number of iterations performed.}
}
\description{
Coordinates the optimization process for the intensity parameters by alternating between Alpha and Beta.
}
