% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nrBeta}
\alias{nrBeta}
\title{Newton-Raphson for Beta coefficients (Fixed Alpha)}
\usage{
nrBeta(
  s,
  n,
  nvar,
  k,
  nk,
  tau,
  caltimes,
  gaptimes,
  censored,
  intercepts,
  slopes,
  lastperrep,
  perrepind,
  effagebegin,
  effage,
  cov,
  alphaSeed,
  betaSeed,
  Z,
  offset,
  rhoFunc,
  ns,
  tol,
  maxiter,
  loglik,
  estiNew,
  infobeta,
  search,
  kiter
)
}
\arguments{
\item{s}{Scale parameter.}

\item{n}{Number of subjects.}

\item{nvar}{Number of covariates.}

\item{k}{Integer vector with the number of events per subject.}

\item{nk}{Total number of observations.}

\item{tau}{Vector of event/censoring times.}

\item{caltimes}{Vector of calendar times.}

\item{gaptimes}{Vector of gap times.}

\item{censored}{Censoring indicator vector.}

\item{intercepts}{Vector of intercepts for the effective age.}

\item{slopes}{Vector of slopes for the effective age.}

\item{lastperrep}{Vector of last perfect repair times.}

\item{perrepind}{Indicator of perfect repair.}

\item{effagebegin}{Effective age at the beginning of the interval.}

\item{effage}{Effective age at the end of the interval.}

\item{cov}{Matrix of covariates.}

\item{alphaSeed}{Initial value for alpha.}

\item{betaSeed}{Initial vector for betas.}

\item{Z}{Vector of frailty (fixed to 1 in this case).}

\item{offset}{Vector of offsets.}

\item{rhoFunc}{Integer indicating the type of rho function.}

\item{ns}{Integer vector for subjects (internal use).}

\item{tol}{Tolerance for convergence.}

\item{maxiter}{Maximum number of iterations.}
}
\value{
A list with: loglik, estim (betas), info (infobeta), search, and kiter.
}
\description{
Optimizes the beta vector using the information matrix.
}
