% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nsm}
\alias{nsm}
\title{Count Events per Subject (nsm)}
\usage{
nsm(s, n, nk, caltimes, k, nm)
}
\arguments{
\item{s}{Current evaluation time.}

\item{n}{Number of subjects.}

\item{nk}{Total number of observations in caltimes.}

\item{caltimes}{Global vector of calendar times.}

\item{k}{Vector with the number of observations for each subject.}

\item{nm}{[Out] Vector to store the results (integer).}
}
\description{
Master routine to fill the event count vector for all subjects.
}
