% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{readmission}
\alias{readmission}
\title{Rehospitalization of colorectal cancer patients}
\format{
A data frame with 861 rows and 10 columns:
\describe{
\item{id}{identifier of each subject, repeated for each recurrence}
\item{enum}{which readmission}
\item{t.start}{start of interval (0 or previous recurrence time)}
\item{t.stop}{recurrence or censoring time}
\item{time}{inter-occurrence or censoring time}
\item{event}{censoring status: 1 for each readmission, 0 for the last
(censored) record of each subject}
\item{chemo}{did the patient receive chemotherapy? 1: No; 2: Yes}
\item{sex}{gender: 1: Males, 2: Females}
\item{dukes}{Dukes' tumoral stage: 1: A-B; 2: C; 3: D}
\item{charlson}{Charlson comorbidity index (time-dependent covariate):
0: index 0; 1: index 1-2; 3: index >= 3}
}
}
\source{
González, J.R., Fernández, E., Moreno, V. et al. (2005). Gender
differences in hospital readmission among colorectal cancer patients.
\emph{Journal of Epidemiology and Community Health}, 59(6), 506--511.
}
\usage{
readmission
}
\description{
Rehospitalization times after surgery in patients diagnosed with
colorectal cancer.
}
\examples{
data(readmission)
head(readmission)
}
\keyword{datasets}
