% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scorefunc}
\alias{scorefunc}
\title{Likelihood and Score Calculation}
\usage{
scorefunc(
  s,
  n,
  nvar,
  k,
  nk,
  tau,
  caltimes,
  gaptimes,
  censored,
  intercepts,
  slopes,
  lastperrep,
  perrepind,
  effagebegin,
  effage,
  cov,
  alpha,
  beta,
  Z,
  offset,
  rhoFunc,
  ns,
  loglik,
  score,
  info,
  scorealpha,
  scorebeta,
  infoalpha,
  infobeta
)
}
\arguments{
\item{s}{Current time point.}

\item{n}{Number of subjects.}

\item{nvar}{Number of covariates.}

\item{k}{Observations per subject.}

\item{nk}{Total number of observations.}

\item{tau}{Censoring times.}

\item{caltimes}{Calendar times.}

\item{gaptimes}{Gap times.}

\item{censored}{Censoring indicators.}

\item{intercepts}{Model intercepts.}

\item{slopes}{Model slopes.}

\item{lastperrep}{Periodic report indices.}

\item{perrepind}{Periodic report indicators.}

\item{effagebegin}{Effective age starts.}

\item{effage}{Effective age ends.}

\item{cov}{Covariate matrix.}

\item{alpha}{Recurrence parameter.}

\item{beta}{Regression coefficients.}

\item{Z}{Fragility/Random effect vector.}

\item{offset}{Offsets vector.}

\item{rhoFunc}{Type of rho function.}

\item{ns}{Vector of event counts per subject.}

\item{loglik}{[Out] Computed log-likelihood.}

\item{score}{[Out] Full score vector (alpha + beta).}

\item{info}{[Out] Full information matrix.}

\item{scorealpha}{[Out] Score component for alpha.}

\item{scorebeta}{[Out] Score component for beta.}

\item{infoalpha}{[Out] Information component for alpha.}

\item{infobeta}{[Out] Information matrix for beta.}
}
\description{
Computes the log-likelihood, score vector, and information matrix 
for a given set of parameters in a recurrent event model.
}
