\name{GeneratedData}
\alias{GeneratedData}
\non_function{}
\title{Simulated data set generated under the minimal repair model}
\usage{data(GeneratedData)}
\description{
 This contains recurrent times under minimal repair model with probability 
 of perfect repair equal to 0.6. Data are as a list (see gcmrec help). 
}
 

\keyword{datasets}
\keyword{survival}