\name{hydraulic}
\alias{hydraulic}
\non_function{}
\title{hydraulic load-haul-dump (LHD) subsystems}
\usage{data(hydraulic)}
\description{
Hydraulic load-haul-dump (LHD) subsystems used in moving
ore and rock in underground mines in Sweden. The data set provides the calendar times
(in hours), excluding repair or down times, of the successive failures of n = 6 such systems
during the two-year development phase.
}

\source{
Kumar, D. and B. Klefsjo (1992). Reliability analysis of hydraulic systems of lhd machines
using the power law process model. \emph{Reliability Engineering and System Safety} 35, 217-
224.

}

\keyword{datasets}
\keyword{survival}
